twitter-ingestion/requirements.txt
tweepy
boto3
python-dotenv